move worker adaptations to `CompositeMap`-renderer
