+++ src/ol/Map.js
@@ -40,7 +40,6 @@ import PointerEventType from './pointer/EventType.js';
 import {fromUserCoordinate, toUserCoordinate} from './proj.js';
 import RenderEventType from './render/EventType.js';
 import CompositeMapRenderer from './renderer/Composite.js';
-import WorkerMapRenderer from './renderer/Worker.js';
 import {hasArea} from './size.js';
 import {
   apply as applyTransform,
@@ -372,7 +371,6 @@ class Map extends BaseObject {
      * @type {!HTMLElement}
      */
     if (!WORKER_OFFSCREEN_CANVAS) {
-      // to do: mock div?
       this.viewport_ = document.createElement('div');
       this.viewport_.className =
         'ol-viewport' + ('ontouchstart' in window ? ' ol-touch' : '');
@@ -1326,9 +1324,7 @@ class Map extends BaseObject {
         targetElement.appendChild(this.viewport_);
       }
       if (!this.renderer_) {
-        this.renderer_ = WORKER_OFFSCREEN_CANVAS
-          ? new WorkerMapRenderer(this)
-          : new CompositeMapRenderer(this);
+        this.renderer_ = new CompositeMapRenderer(this);
       }
 
       if (!WORKER_OFFSCREEN_CANVAS) {
